Added back base test class, and added new class for raw web socket tests
