+++ test/Microsoft.Azure.Relay.UnitTests/Constants.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Azure.Relay.UnitTests
+{
+    static class Constants
+    {
+        internal const string ConnectionStringEnvironmentVariable = "azure-relay-dotnet/connectionstring";
+        internal const string AuthenticatedEntityPath = "authenticated";
+        internal const string UnauthenticatedEntityPath = "unauthenticated";
+    }
+}
