+++ test/Microsoft.Azure.Relay.UnitTests/RunTimeTests.cs
@@ -5,49 +5,14 @@ namespace Microsoft.Azure.Relay.UnitTests
 {
     using System;
     using System.Collections.Generic;
-    using System.IO;
-    using System.Net;
-    using System.Net.WebSockets;
     using System.Threading;
     using System.Threading.Tasks;
     using Xunit;
 
-    public class RunTimeTests
+    public class RunTimeTests : HybridConnectionTestBase
     {
-        const string ConnectionStringEnvironmentVariable = "azure-relay-dotnet/connectionstring";
-        const string AuthenticatedEntityPath = "authenticated";
-        const string UnauthenticatedEntityPath = "unauthenticated";
-
-        const string TestAcceptHandlerResultHeader = "X-TestAcceptHandlerResult";
-        const string TestAcceptHandlerStatusCodeHeader = "X-TestAcceptHandlerStatusCode";
-        const string TestAcceptHandlerStatusDescriptionHeader = "X-TestAcceptHandlerStatusDescription";
-        const string TestAcceptHandlerSetResponseHeader = "X-TestAcceptHandlerSetResponseHeader";
-        const string TestAcceptHandlerDelayHeader = "X-TestAcceptHandlerDelay";
-
-        readonly string ConnectionString;
-
-        enum EndpointTestType
+        public RunTimeTests() : base()
         {
-            Authenticated,
-            Unauthenticated
-        }
-
-        public static IEnumerable<object> AuthenticationTestPermutations => new object[]
-        {
-            new object[] { EndpointTestType.Authenticated },
-            new object[] { EndpointTestType.Unauthenticated }
-        };
-
-        public RunTimeTests()
-        {
-            var envConnectionString = Environment.GetEnvironmentVariable(ConnectionStringEnvironmentVariable);
-            if (string.IsNullOrWhiteSpace(envConnectionString))
-            {
-                throw new InvalidOperationException($"'{ConnectionStringEnvironmentVariable}' environment variable was not found!");
-            }
-
-            // Validate the connection string
-            ConnectionString = new RelayConnectionStringBuilder(envConnectionString).ToString();
         }
 
         [Theory, DisplayTestMethodName]
@@ -57,7 +22,7 @@ namespace Microsoft.Azure.Relay.UnitTests
             HybridConnectionListener listener = null;
             try
             {
-                listener = GetHybridConnectionListener(endpointTestType);
+                listener = this.GetHybridConnectionListener(endpointTestType);
                 var client = GetHybridConnectionClient(endpointTestType);
 
                 TestUtility.Log($"Opening {listener}");
@@ -344,11 +309,11 @@ namespace Microsoft.Azure.Relay.UnitTests
 
                 if (endpointTestType == EndpointTestType.Authenticated)
                 {
-                    fakeEndpointConnectionStringBuilder.EntityPath = AuthenticatedEntityPath;
+                    fakeEndpointConnectionStringBuilder.EntityPath = Constants.AuthenticatedEntityPath;
                 }
                 else
                 {
-                    fakeEndpointConnectionStringBuilder.EntityPath = UnauthenticatedEntityPath;
+                    fakeEndpointConnectionStringBuilder.EntityPath = Constants.UnauthenticatedEntityPath;
                 }
 
                 var fakeEndpointConnectionString = fakeEndpointConnectionStringBuilder.ToString();
@@ -437,220 +402,6 @@ namespace Microsoft.Azure.Relay.UnitTests
             }
         }
 
-        [Theory, DisplayTestMethodName]
-        [MemberData(nameof(AuthenticationTestPermutations))]
-        async Task RawWebSocketSenderTest(EndpointTestType endpointTestType)
-        {
-            HybridConnectionListener listener = null;
-            try
-            {
-                listener = GetHybridConnectionListener(endpointTestType);
-
-                var clientWebSocket = new ClientWebSocket();
-                var wssUri = await GetWebSocketConnectionUri(endpointTestType);
-
-                using (var cancelSource = new CancellationTokenSource(TimeSpan.FromMinutes(1)))
-                {
-                    TestUtility.Log($"Opening {listener}");
-                    await listener.OpenAsync(cancelSource.Token);
-
-                    await clientWebSocket.ConnectAsync(wssUri, cancelSource.Token);
-
-                    var listenerStream = await listener.AcceptConnectionAsync();
-
-                    TestUtility.Log("Client and Listener are connected!");
-                    Assert.Null(clientWebSocket.SubProtocol);
-
-                    byte[] sendBuffer = this.CreateBuffer(1024, new byte[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 });
-                    await clientWebSocket.SendAsync(new ArraySegment<byte>(sendBuffer), WebSocketMessageType.Binary, true, cancelSource.Token);
-                    TestUtility.Log($"clientWebSocket wrote {sendBuffer.Length} bytes");
-
-                    byte[] readBuffer = new byte[sendBuffer.Length];
-                    await this.ReadCountBytesAsync(listenerStream, readBuffer, 0, readBuffer.Length, TimeSpan.FromSeconds(30));
-                    Assert.Equal(sendBuffer, readBuffer);
-
-                    TestUtility.Log("Calling clientStream.CloseAsync");
-                    var clientStreamCloseTask = clientWebSocket.CloseAsync(WebSocketCloseStatus.EndpointUnavailable, "From Test Code", cancelSource.Token);
-                    TestUtility.Log("Reading from listenerStream");
-                    int bytesRead = await this.SafeReadAsync(listenerStream, readBuffer, 0, readBuffer.Length);
-                    TestUtility.Log($"listenerStream.Read returned {bytesRead} bytes");
-                    Assert.Equal(0, bytesRead);
-
-                    TestUtility.Log("Calling listenerStream.CloseAsync");
-                    var listenerStreamCloseTask = listenerStream.CloseAsync(cancelSource.Token);
-                    await listenerStreamCloseTask;
-                    TestUtility.Log("Calling listenerStream.CloseAsync completed");
-                    await clientStreamCloseTask;
-                    TestUtility.Log("Calling clientStream.CloseAsync completed");
-
-                    TestUtility.Log($"Closing {listener}");
-                    await listener.CloseAsync(TimeSpan.FromSeconds(10));
-                    listener = null;
-                }
-            }
-            finally
-            {
-                await this.SafeCloseAsync(listener);
-            }
-        }
-
-        [Theory, DisplayTestMethodName]
-        [MemberData(nameof(AuthenticationTestPermutations))]
-        async Task AcceptHandlerTest(EndpointTestType endpointTestType)
-        {
-            HybridConnectionListener listener = null;
-            try
-            {
-                listener = GetHybridConnectionListener(endpointTestType);
-                TestUtility.Log($"Opening {listener}");
-                await listener.OpenAsync(TimeSpan.FromSeconds(30));
-
-                // Install a Custom AcceptHandler which allows using Headers to control the StatusCode, StatusDescription,
-                // and whether to accept or reject the client.
-                listener.AcceptHandler = TestAcceptHandler;
-
-                var clientConnectionString = GetConnectionString(endpointTestType);
-
-                var connectionStringBuilder = new RelayConnectionStringBuilder(clientConnectionString);
-
-                var wssUri = await GetWebSocketConnectionUri(endpointTestType);
-                TestUtility.Log($"Using WebSocket address {wssUri}");
-                using (var cancelSource = new CancellationTokenSource(TimeSpan.FromMinutes(1)))
-                {
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler accepting with only returning true");
-                    AcceptEchoListener(listener);
-                    var clientWebSocket = new ClientWebSocket();
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerResultHeader, true.ToString());
-                    await clientWebSocket.ConnectAsync(wssUri, cancelSource.Token);
-                    await clientWebSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Test closing web socket", cancelSource.Token);
-
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler accepting and setting response header");
-                    AcceptEchoListener(listener);
-                    clientWebSocket = new ClientWebSocket();
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerResultHeader, true.ToString());
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerSetResponseHeader, "X-CustomHeader: Test value");
-                    await clientWebSocket.ConnectAsync(wssUri, cancelSource.Token);
-                    await clientWebSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Test closing web socket", cancelSource.Token);
-
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler rejecting with only returning false");
-                    AcceptEchoListener(listener);
-                    HttpStatusCode expectedStatusCode = HttpStatusCode.BadRequest;
-                    string expectedStatusDescription = "Rejected by user code";
-                    clientWebSocket = new ClientWebSocket();
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerResultHeader, false.ToString());
-                    var webSocketException = await Assert.ThrowsAsync<WebSocketException>(() => clientWebSocket.ConnectAsync(wssUri, cancelSource.Token));
-                    VerifyHttpStatusCodeAndDescription(webSocketException, expectedStatusCode, expectedStatusDescription);
-
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler rejecting with setting status code and returning false");
-                    AcceptEchoListener(listener);
-                    expectedStatusCode = HttpStatusCode.Unauthorized;
-                    expectedStatusDescription = "Unauthorized";
-                    clientWebSocket = new ClientWebSocket();
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerResultHeader, false.ToString());
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerStatusCodeHeader, ((int)expectedStatusCode).ToString());
-                    webSocketException = await Assert.ThrowsAsync<WebSocketException>(() => clientWebSocket.ConnectAsync(wssUri, cancelSource.Token));
-                    VerifyHttpStatusCodeAndDescription(webSocketException, expectedStatusCode, expectedStatusDescription);
-
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler rejecting with setting status code+description and returning false");
-                    AcceptEchoListener(listener);
-                    expectedStatusCode = HttpStatusCode.Unauthorized;
-                    expectedStatusDescription = "Status Description from test";
-                    clientWebSocket = new ClientWebSocket();
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerResultHeader, false.ToString());
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerStatusCodeHeader, ((int)expectedStatusCode).ToString());
-                    clientWebSocket.Options.SetRequestHeader(TestAcceptHandlerStatusDescriptionHeader, expectedStatusDescription);
-                    webSocketException = await Assert.ThrowsAsync<WebSocketException>(() => clientWebSocket.ConnectAsync(wssUri, cancelSource.Token));
-                    VerifyHttpStatusCodeAndDescription(webSocketException, expectedStatusCode, expectedStatusDescription);
-
-                    TestUtility.Log("Testing HybridConnectionListener.AcceptHandler with a custom handler that returns null instead of a valid Task<bool>");
-                    listener.AcceptHandler = context => null;
-                    AcceptEchoListener(listener);
-                    expectedStatusCode = HttpStatusCode.BadGateway;
-                    expectedStatusDescription = "The Listener's custom AcceptHandler threw an exception. See Listener logs for details.";
-                    clientWebSocket = new ClientWebSocket();
-                    webSocketException = await Assert.ThrowsAsync<WebSocketException>(() => clientWebSocket.ConnectAsync(wssUri, cancelSource.Token));
-                    VerifyHttpStatusCodeAndDescription(webSocketException, expectedStatusCode, expectedStatusDescription, false);
-                }
-            }
-            finally
-            {
-                await SafeCloseAsync(listener);
-            }
-        }
-
-        /// <summary>
-        /// Since these tests all share a common connection string, this method will modify the 
-        /// endpoint / shared access keys as needed based on the EndpointTestType.
-        /// </summary>
-        string GetConnectionString(EndpointTestType endpointTestType)
-        {
-            var connectionStringBuilder = new RelayConnectionStringBuilder(this.ConnectionString);
-            if (endpointTestType == EndpointTestType.Unauthenticated)
-            {
-                connectionStringBuilder.EntityPath = UnauthenticatedEntityPath;
-                connectionStringBuilder.SharedAccessKey = string.Empty;
-                connectionStringBuilder.SharedAccessKeyName = string.Empty;
-                connectionStringBuilder.SharedAccessSignature = string.Empty;
-            }
-            else
-            {
-                connectionStringBuilder.EntityPath = AuthenticatedEntityPath;
-            }
-
-            return connectionStringBuilder.ToString();
-        }
-
-        /// <summary>
-        /// Returns a HybridConnectionClient based on the EndpointTestType (authenticated/unauthenticated).
-        /// </summary>
-        HybridConnectionClient GetHybridConnectionClient(EndpointTestType endpointTestType)
-        {
-            var connectionString = GetConnectionString(endpointTestType);
-            return new HybridConnectionClient(connectionString);
-        }
-
-        /// <summary>
-        /// Returns a HybridConnectionListener based on the EndpointTestType (authenticated/unauthenticated).
-        /// </summary>
-        HybridConnectionListener GetHybridConnectionListener(EndpointTestType endpointTestType)
-        {
-            // Even if the endpoint is unauthenticated, the *listener* still needs to be authenticated
-            if (endpointTestType == EndpointTestType.Unauthenticated)
-            {
-                var connectionStringBuilder = new RelayConnectionStringBuilder(this.ConnectionString)
-                {
-                    EntityPath = UnauthenticatedEntityPath
-                };
-                return new HybridConnectionListener(connectionStringBuilder.ToString());
-            }
-
-            return new HybridConnectionListener(GetConnectionString(endpointTestType));
-        }
-
-        /// <summary>
-        /// Since these tests all share a common connection string, this method will modify the 
-        /// endpoint / shared access keys as needed based on the EndpointTestType, and return a WebSocket URI.
-        /// </summary>
-        async Task<Uri> GetWebSocketConnectionUri(EndpointTestType endpointTestType)
-        {
-            var clientConnectionString = GetConnectionString(endpointTestType);
-            var connectionStringBuilder = new RelayConnectionStringBuilder(clientConnectionString);
-            var connectionUriString = $"wss://{connectionStringBuilder.Endpoint.Host}/$hc/{connectionStringBuilder.EntityPath}";
-
-            if (endpointTestType == EndpointTestType.Authenticated)
-            {
-                var tokenProvider = TokenProvider.CreateSharedAccessSignatureTokenProvider(
-                    connectionStringBuilder.SharedAccessKeyName,
-                    connectionStringBuilder.SharedAccessKey);
-
-                var token = await tokenProvider.GetTokenAsync(connectionStringBuilder.Endpoint.ToString(), TimeSpan.FromMinutes(10));
-
-                connectionUriString += $"?sb-hc-token={WebUtility.UrlEncode(token.TokenString)}";
-            }
-
-            return new Uri(connectionUriString);
-        }
-
         /// <summary>
         /// Create an send-side HybridConnectionStream, send N bytes to it, receive N bytes response from it,
         /// then close the HybridConnectionStream.
@@ -679,219 +430,6 @@ namespace Microsoft.Azure.Relay.UnitTests
             {
                 cancelSource.Dispose();
             }
-        }
-
-        /// <summary>
-        /// Call HybridConnectionListener.AcceptConnectionAsync, once/if a listener is accepted
-        /// read from its stream and echo the bytes until a 0-byte read occurs, then close.
-        /// </summary>
-        async void AcceptEchoListener(HybridConnectionListener listener)
-        {
-            try
-            {
-                var listenerStream = await listener.AcceptConnectionAsync();
-                if (listenerStream != null)
-                {
-                    byte[] buffer = new byte[4 * 1024];
-                    do
-                    {
-                        int bytesRead;
-                        try
-                        {
-                            bytesRead = await listenerStream.ReadAsync(buffer, 0, buffer.Length);
-                        }
-                        catch (Exception readException)
-                        {
-                            TestUtility.Log($"AcceptEchoListener {readException.GetType().Name}: {readException.Message}");
-                            using (var cts = new CancellationTokenSource(TimeSpan.FromSeconds(10)))
-                            {
-                                await listenerStream.CloseAsync(cts.Token);
-                            }
-
-                            return;
-                        }
-
-                        using (var cts = new CancellationTokenSource(TimeSpan.FromSeconds(10)))
-                        {
-                            if (bytesRead == 0)
-                            {
-                                await listenerStream.CloseAsync(cts.Token);
-                                return;
-                            }
-                            else
-                            {
-                                await listenerStream.WriteAsync(buffer, 0, bytesRead, cts.Token);
-                            }
-                        }
-                    }
-                    while (true);
-                }
-            }
-            catch (Exception e)
-            {
-                TestUtility.Log($"AcceptEchoListener {e.GetType().Name}: {e.Message}");
-            }
-        }
-
-        async Task ReadCountBytesAsync(Stream stream, byte[] buffer, int offset, int bytesToRead, TimeSpan timeout)
-        {
-            DateTime timeoutInstant = DateTime.Now.Add(timeout);
-            int? originalReadTimeout = stream.CanTimeout ? stream.ReadTimeout : (int?)null;
-            try
-            {
-                int totalBytesRead = 0;
-                do
-                {
-                    TimeSpan remainingTimeout = timeoutInstant.Subtract(DateTime.Now);
-                    if (remainingTimeout <= TimeSpan.Zero)
-                    {
-                        break;
-                    }
-
-                    stream.ReadTimeout = (int)remainingTimeout.TotalMilliseconds;
-                    int bytesRead = await stream.ReadAsync(buffer, offset + totalBytesRead, bytesToRead - totalBytesRead);
-                    TestUtility.Log($"Stream read {bytesRead} bytes");
-                    if (bytesRead == 0)
-                    {
-                        break;
-                    }
-
-                    totalBytesRead += bytesRead;
-                }
-                while (totalBytesRead < bytesToRead);
-
-                if (totalBytesRead < bytesToRead)
-                {
-                    throw new TimeoutException("The requested number of bytes could not be received.  ReadAsync returned 0 bytes");
-                }
-            }
-            finally
-            {
-                if (originalReadTimeout.HasValue)
-                {
-                    stream.ReadTimeout = originalReadTimeout.Value;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Calls Stream.ReadAsync with Exception handling. If an IOException occurs a zero byte read is returned.
-        /// </summary>
-        async Task<int> SafeReadAsync(Stream stream, byte[] buffer, int offset, int bytesToRead)
-        {
-            int bytesRead = 0;
-            try
-            {
-                bytesRead = await stream.ReadAsync(buffer, offset, bytesToRead);
-                TestUtility.Log($"Stream read {bytesRead} bytes");
-            }
-            catch (IOException ex)
-            {
-                TestUtility.Log($"Stream.ReadAsync error {ex}");
-            }
-
-            return bytesRead;
-        }
-
-        byte[] CreateBuffer(int length, byte[] fillPattern)
-        {
-            byte[] buffer = new byte[length];
-
-            int offset = 0;
-            do
-            {
-                int bytesToCopy = Math.Min(length - offset, fillPattern.Length);
-                Array.Copy(fillPattern, 0, buffer, offset, bytesToCopy);
-                offset += bytesToCopy;
-            }
-            while (offset < length);
-
-            return buffer;
-        }
-
-        async Task SafeCloseAsync(HybridConnectionListener listener)
-        {
-            if (listener != null)
-            {
-                try
-                {
-                    TestUtility.Log($"Closing {listener}");
-                    await listener.CloseAsync(TimeSpan.FromSeconds(10));
-                }
-                catch (Exception e)
-                {
-                    TestUtility.Log($"Error closing HybridConnectionListener {e.GetType()}: {e.Message}");
-                }
-            }
-        }
-
-        async Task<bool> TestAcceptHandler(RelayedHttpListenerContext listenerContext)
-        {
-            string delayString = listenerContext.Request.Headers[TestAcceptHandlerDelayHeader];
-            TimeSpan delay;
-            if (!string.IsNullOrEmpty(delayString) && TimeSpan.TryParse(delayString, out delay))
-            {
-                await Task.Delay(delay);
-            }
-
-            string statusCodeString = listenerContext.Request.Headers[TestAcceptHandlerStatusCodeHeader];
-            int statusCode;
-            if (!string.IsNullOrEmpty(statusCodeString) && int.TryParse(statusCodeString, out statusCode))
-            {
-                listenerContext.Response.StatusCode = (HttpStatusCode)statusCode;
-            }
-
-            string statusDescription = listenerContext.Request.Headers[TestAcceptHandlerStatusDescriptionHeader];
-            if (!string.IsNullOrEmpty(statusDescription))
-            {
-                listenerContext.Response.StatusDescription = statusDescription;
-            }
-
-            string responseHeaders = listenerContext.Request.Headers[TestAcceptHandlerSetResponseHeader];
-            if (!string.IsNullOrEmpty(responseHeaders))
-            {
-                foreach (var headerAndValue in responseHeaders.Split(','))
-                {
-                    string[] headerNameAndValue = headerAndValue.Split(':');
-                    listenerContext.Response.Headers[headerNameAndValue[0]] = headerNameAndValue[1].Trim();
-                }
-            }
-
-            bool result;
-            if (!bool.TryParse(listenerContext.Request.Headers[TestAcceptHandlerResultHeader], out result))
-            {
-                result = true;
-            }
-
-            TestUtility.Log($"Test AcceptHandler: {listenerContext.Request.Url} {(int)listenerContext.Response.StatusCode}: {listenerContext.Response.StatusDescription} returning {result}");
-            return result;
-        }
-
-        void VerifyHttpStatusCodeAndDescription(
-            WebSocketException webSocketException,
-            HttpStatusCode expectedStatusCode,
-            string expectedStatusDescription,
-            bool exactMatchDescription = true)
-        {
-            Assert.NotNull(webSocketException.InnerException);
-
-            // TODO: Error details aren't available in .NET Core due to issue:
-            // https://github.com/dotnet/corefx/issues/13773
-            ////Assert.IsAssignableFrom<WebException>(webSocketException.InnerException);
-            ////var webException = (WebException)webSocketException.InnerException;
-            ////Assert.NotNull(webException.Response);
-            ////Assert.IsAssignableFrom<HttpWebResponse>(webException.Response);
-            ////var httpWebResponse = (HttpWebResponse)webException.Response;
-            ////TestUtility.Log($"Actual HTTP Status: {(int)httpWebResponse.StatusCode}: {httpWebResponse.StatusDescription}");
-            ////Assert.Equal(expectedStatusCode, httpWebResponse.StatusCode);
-            ////if (exactMatchDescription)
-            ////{
-            ////    Assert.Equal(expectedStatusDescription, httpWebResponse.StatusDescription);
-            ////}
-            ////else
-            ////{
-            ////    Assert.Contains(expectedStatusDescription, httpWebResponse.StatusDescription);
-            ////}
         }
     }
 }
